front: move parseFile helper to import folder
