+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/helpers/parseJson.ts
